Refactor code to remove fileam.h including from copy.
Remove the fileam.h including from non external code. So we can extract
external table into extension.
Implement external table fdw's BeginForeignInsert and EndForeignInsert,
so COPY FROM will go throug the fdw routine instead of the hack
external insert.
