+++ src/backend/executor/nodeModifyTable.c
@@ -53,7 +53,6 @@
 #include "utils/rel.h"
 #include "utils/tqual.h"
 
-#include "access/fileam.h"
 #include "access/transam.h"
 #include "cdb/cdbaocsam.h"
 #include "cdb/cdbappendonlyam.h"
