+++ src/include/access/fileam.h
@@ -84,13 +84,6 @@ external_getnext(FileScanDesc scan,
 extern ExternalInsertDesc external_insert_init(Relation rel);
 extern Oid	external_insert(ExternalInsertDesc extInsertDesc, HeapTuple instup);
 extern void external_insert_finish(ExternalInsertDesc extInsertDesc);
-extern void external_set_env_vars(extvar_t *extvar, char *uri, bool csv, char *escape, char *quote, bool header, uint32 scancounter);
-extern char *linenumber_atoi(char *buffer, size_t bufsz, int64 linenumber);
-
-/* prototypes for functions in url_execute.c */
-extern int popen_with_stderr(int *rwepipe, const char *exe, bool forwrite);
-extern int pclose_with_stderr(int pid, int *rwepipe, StringInfo sinfo);
-extern char *make_command(const char *cmd, extvar_t *ev);
 
 extern List *appendCopyEncodingOption(List *copyFmtOpts, int encoding);
