+++ src/include/utils/formatting.h
@@ -27,6 +27,7 @@ extern char *str_initcap(const char *buff, size_t nbytes, Oid collid);
 extern char *asc_tolower(const char *buff, size_t nbytes);
 extern char *asc_toupper(const char *buff, size_t nbytes);
 extern char *asc_initcap(const char *buff, size_t nbytes);
+extern char *linenumber_atoi(char *buffer, size_t bufsz, int64 linenumber);
 
 extern Datum timestamp_to_char(PG_FUNCTION_ARGS);
 extern Datum timestamptz_to_char(PG_FUNCTION_ARGS);
